:tada: 4.8.0.RELEASE 新增接口权限模块，增强系统安全性

+++ src/main/java/org/springblade/modules/system/service/impl/RoleServiceImpl.java
@@ -40,6 +40,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static org.springblade.common.constant.CommonConstant.API_SCOPE_CATEGORY;
+import static org.springblade.common.constant.CommonConstant.DATA_SCOPE_CATEGORY;
+
 /**
  * 服务实现类
  *
@@ -69,7 +72,11 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
 	}
 
 	@Override
-	public boolean grant(@NotEmpty List<Long> roleIds, @NotEmpty List<Long> menuIds, List<Long> dataScopeIds) {
+	public boolean grant(@NotEmpty List<Long> roleIds, List<Long> menuIds, List<Long> dataScopeIds, List<Long> apiScopeIds) {
+		return grantRoleMenu(roleIds, menuIds) && grantDataScope(roleIds, dataScopeIds) && grantApiScope(roleIds, apiScopeIds);
+	}
+
+	private boolean grantRoleMenu(List<Long> roleIds, List<Long> menuIds) {
 		// 删除角色配置的菜单集合
 		roleMenuService.remove(Wrappers.<RoleMenu>update().lambda().in(RoleMenu::getRoleId, roleIds));
 		// 组装配置
@@ -82,28 +89,47 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
 		}));
 		// 新增配置
 		roleMenuService.saveBatch(roleMenus);
+		return true;
+	}
 
+	private boolean grantDataScope(List<Long> roleIds, List<Long> dataScopeIds) {
 		// 删除角色配置的数据权限集合
-		roleScopeService.remove(Wrappers.<RoleScope>update().lambda().in(RoleScope::getRoleId, roleIds));
+		roleScopeService.remove(Wrappers.<RoleScope>update().lambda().eq(RoleScope::getScopeCategory, DATA_SCOPE_CATEGORY).in(RoleScope::getRoleId, roleIds));
 		// 组装配置
 		List<RoleScope> roleDataScopes = new ArrayList<>();
 		roleIds.forEach(roleId -> dataScopeIds.forEach(scopeId -> {
 			RoleScope roleScope = new RoleScope();
+			roleScope.setScopeCategory(DATA_SCOPE_CATEGORY);
 			roleScope.setRoleId(roleId);
 			roleScope.setScopeId(scopeId);
 			roleDataScopes.add(roleScope);
 		}));
 		// 新增配置
 		roleScopeService.saveBatch(roleDataScopes);
-
 		return true;
+	}
 
+	private boolean grantApiScope(List<Long> roleIds, List<Long> apiScopeIds) {
+		// 删除角色配置的接口权限集合
+		roleScopeService.remove(Wrappers.<RoleScope>update().lambda().eq(RoleScope::getScopeCategory, API_SCOPE_CATEGORY).in(RoleScope::getRoleId, roleIds));
+		// 组装配置
+		List<RoleScope> roleApiScopes = new ArrayList<>();
+		roleIds.forEach(roleId -> apiScopeIds.forEach(scopeId -> {
+			RoleScope roleScope = new RoleScope();
+			roleScope.setScopeCategory(API_SCOPE_CATEGORY);
+			roleScope.setScopeId(scopeId);
+			roleScope.setRoleId(roleId);
+			roleApiScopes.add(roleScope);
+		}));
+		// 新增配置
+		roleScopeService.saveBatch(roleApiScopes);
+		return true;
 	}
 
 	@Override
 	public String getRoleIds(String tenantId, String roleNames) {
 		List<Role> roleList = baseMapper.selectList(Wrappers.<Role>query().lambda().eq(Role::getTenantId, tenantId).in(Role::getRoleName, Func.toStrList(roleNames)));
-		if (roleList != null && roleList.size() > 0) {
+		if (roleList != null && !roleList.isEmpty()) {
 			return roleList.stream().map(role -> Func.toStr(role.getId())).distinct().collect(Collectors.joining(","));
 		}
 		return null;
